Spin off binding data into BindUtils
This is in preparation for doing some very similar things from the PCA script
